+++ asyncgit/src/sync/commit_files.rs
@@ -8,7 +8,37 @@ use crate::{
 };
 use git2::{Diff, Repository};
 use scopetime::scope_time;
-use std::{cmp::Ordering, collections::HashSet};
+use std::collections::HashSet;
+
+/// struct containing a new and an old version
+#[derive(Copy, Clone, PartialEq, Eq, Hash, Debug)]
+pub struct OldNew<T> {
+	/// The old version
+	pub old: T,
+	/// The new version
+	pub new: T,
+}
+
+/// Sort two commits.
+pub fn sort_commits(
+	repo: &Repository,
+	commits: (CommitId, CommitId),
+) -> Result<OldNew<CommitId>> {
+	if repo.graph_descendant_of(
+		commits.0.get_oid(),
+		commits.1.get_oid(),
+	)? {
+		Ok(OldNew {
+			old: commits.1,
+			new: commits.0,
+		})
+	} else {
+		Ok(OldNew {
+			old: commits.0,
+			new: commits.1,
+		})
+	}
+}
 
 /// get all files that are part of a commit
 pub fn get_commit_files(
@@ -21,7 +51,12 @@ pub fn get_commit_files(
 	let repo = repo(repo_path)?;
 
 	let diff = if let Some(other) = other {
-		get_compare_commits_diff(&repo, (id, other), None, None)?
+		get_compare_commits_diff(
+			&repo,
+			sort_commits(&repo, (id, other))?,
+			None,
+			None,
+		)?
 	} else {
 		get_commit_diff(
 			&repo,
@@ -55,26 +90,20 @@ pub fn get_commit_files(
 #[allow(clippy::needless_pass_by_value)]
 pub fn get_compare_commits_diff(
 	repo: &Repository,
-	ids: (CommitId, CommitId),
+	ids: OldNew<CommitId>,
 	pathspec: Option<String>,
 	options: Option<DiffOptions>,
 ) -> Result<Diff<'_>> {
 	// scope_time!("get_compare_commits_diff");
-
-	let commits = (
-		repo.find_commit(ids.0.into())?,
-		repo.find_commit(ids.1.into())?,
-	);
-
-	let commits = if commits.0.time().cmp(&commits.1.time())
-		== Ordering::Greater
-	{
-		(commits.1, commits.0)
-	} else {
-		commits
+	let commits = OldNew {
+		old: repo.find_commit(ids.old.into())?,
+		new: repo.find_commit(ids.new.into())?,
 	};
 
-	let trees = (commits.0.tree()?, commits.1.tree()?);
+	let trees = OldNew {
+		old: commits.old.tree()?,
+		new: commits.new.tree()?,
+	};
 
 	let mut opts = git2::DiffOptions::new();
 	if let Some(options) = options {
@@ -86,9 +115,9 @@ pub fn get_compare_commits_diff(
 		opts.pathspec(p.clone());
 	}
 
-	let diff = repo.diff_tree_to_tree(
-		Some(&trees.0),
-		Some(&trees.1),
+	let diff: Diff<'_> = repo.diff_tree_to_tree(
+		Some(&trees.old),
+		Some(&trees.new),
 		Some(&mut opts),
 	)?;