+++ src/components/compare_commits.rs
@@ -5,19 +5,18 @@ use super::{
 };
 use crate::{
 	accessors,
+	app::Environment,
 	keys::{key_match, SharedKeyConfig},
 	options::SharedOptions,
 	queue::{InternalEvent, Queue, StackablePopupOpen},
 	strings,
-	ui::style::SharedTheme,
 };
 use anyhow::Result;
 use asyncgit::{
-	sync::{self, CommitId, RepoPathRef},
+	sync::{self, commit_files::OldNew, CommitId, RepoPathRef},
 	AsyncDiff, AsyncGitNotification, CommitFilesParams, DiffParams,
 	DiffType,
 };
-use crossbeam_channel::Sender;
 use crossterm::event::Event;
 use ratatui::{
 	backend::Backend,
@@ -168,37 +167,20 @@ impl CompareCommitsComponent {
 	accessors!(self, [diff, details]);
 
 	///
-	pub fn new(
-		repo: &RepoPathRef,
-		queue: &Queue,
-		sender: &Sender<AsyncGitNotification>,
-		theme: SharedTheme,
-		key_config: SharedKeyConfig,
-		options: SharedOptions,
-	) -> Self {
+	pub fn new(env: &Environment) -> Self {
 		Self {
-			repo: repo.clone(),
-			details: CommitDetailsComponent::new(
-				repo,
-				queue,
-				sender,
-				theme.clone(),
-				key_config.clone(),
-			),
-			diff: DiffComponent::new(
-				repo.clone(),
-				queue.clone(),
-				theme,
-				key_config.clone(),
-				true,
-				options.clone(),
-			),
+			repo: env.repo.clone(),
+			details: CommitDetailsComponent::new(env),
+			diff: DiffComponent::new(env, true),
 			open_request: None,
-			git_diff: AsyncDiff::new(repo.borrow().clone(), sender),
+			git_diff: AsyncDiff::new(
+				env.repo.borrow().clone(),
+				&env.sender_git,
+			),
 			visible: false,
-			key_config,
-			queue: queue.clone(),
-			options,
+			key_config: env.key_config.clone(),
+			queue: env.queue.clone(),
+			options: env.options.clone(),
 		}
 	}
 
@@ -240,16 +222,19 @@ impl CompareCommitsComponent {
 		Ok(())
 	}
 
-	fn get_ids(&self) -> Option<(CommitId, CommitId)> {
+	fn get_ids(&self) -> Option<OldNew<CommitId>> {
 		let other = self
 			.open_request
 			.as_ref()
 			.and_then(|open| open.compare_id);
 
-		self.open_request
-			.as_ref()
-			.map(|open| open.commit_id)
-			.zip(other)
+		let this =
+			self.open_request.as_ref().map(|open| open.commit_id);
+
+		Some(OldNew {
+			old: other?,
+			new: this?,
+		})
 	}
 
 	/// called when any tree component changed selection