+++ src/components/cred.rs
@@ -4,6 +4,7 @@ use ratatui::{backend::Backend, layout::Rect, Frame};
 
 use asyncgit::sync::cred::BasicAuthCredential;
 
+use crate::app::Environment;
 use crate::components::{EventState, InputType, TextInputComponent};
 use crate::keys::key_match;
 use crate::{
@@ -13,7 +14,6 @@ use crate::{
 	},
 	keys::SharedKeyConfig,
 	strings,
-	ui::style::SharedTheme,
 };
 
 ///
@@ -27,23 +27,19 @@ pub struct CredComponent {
 
 impl CredComponent {
 	///
-	pub fn new(
-		theme: SharedTheme,
-		key_config: SharedKeyConfig,
-	) -> Self {
+	pub fn new(env: &Environment) -> Self {
+		let key_config = env.key_config.clone();
 		Self {
 			visible: false,
 			input_username: TextInputComponent::new(
-				theme.clone(),
-				key_config.clone(),
+				env,
 				&strings::username_popup_title(&key_config),
 				&strings::username_popup_msg(&key_config),
 				false,
 			)
 			.with_input_type(InputType::Singleline),
 			input_password: TextInputComponent::new(
-				theme,
-				key_config.clone(),
+				env,
 				&strings::password_popup_title(&key_config),
 				&strings::password_popup_msg(&key_config),
 				false,